Moved sub configs to `.enb/config`
