+++ .enb/make.js
@@ -22,12 +22,12 @@ var BEM_TEMPLATE_ENGINE = process.env.BEM_TEMPLATE_ENGINE || 'BH',
     mergeFiles = require('enb/techs/file-merge'),
     mergeBemdecl = require('enb-bem-techs/techs/merge-bemdecl'),
     borschik = require('enb-borschik/techs/borschik'),
-    libLevels = require('./levels'),
-    libs = require('./libs'),
-    langs = require('./langs'),
-    browsers = require('./browsers'),
-    sets = require('./sets'),
-    platforms = require('./platforms'),
+    libLevels = require('./config/levels'),
+    libs = require('./config/libs'),
+    langs = require('./config/langs'),
+    browsers = require('./config/browsers'),
+    sets = require('./config/sets'),
+    platforms = require('./config/platforms'),
     distPlatforms = ['desktop', 'touch-pad', 'touch-phone'];
 
 module.exports = function(config) {
